pull all fluid slot thing to GT_Container
