+++ src/main/java/gregtech/api/gui/GT_Container_BasicMachine.java
@@ -2,6 +2,7 @@ package gregtech.api.gui;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import gregtech.api.interfaces.IFluidAccess;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.metatileentity.implementations.GT_MetaTileEntity_BasicMachine;
 import gregtech.api.metatileentity.implementations.GT_MetaTileEntity_BasicTank;
@@ -263,7 +264,7 @@ public class GT_Container_BasicMachine extends GT_Container_BasicTank {
                         tTank.setFillableStack(GT_Utility.getFluidFromDisplayStack(tTank.getStackInSlot(2)));
                     }
                     GT_MetaTileEntity_BasicTank tTank = (GT_MetaTileEntity_BasicTank) mTileEntity.getMetaTileEntity();
-                    IFluidAccess tFillableAccess = IFluidAccess.from(tTank, true);
+                    BasicTankFluidAccess tFillableAccess = BasicTankFluidAccess.from(tTank, true);
                     GT_Recipe_Map recipes = machine.getRecipeList();
                     //If the  machine has recipes but no fluid inputs, disallow filling this slot with fluids.
                     ItemStack tToken = handleFluidSlotClick(tFillableAccess, aPlayer, aMouseclick == 0, true, (recipes == null || recipes.hasFluidInputs()));
