+++ src/main/java/gregtech/api/interfaces/IFluidAccess.java
@@ -1,0 +1,12 @@
+package gregtech.api.interfaces;
+
+import net.minecraftforge.fluids.FluidStack;
+
+public interface IFluidAccess {
+
+    void set(FluidStack stack);
+
+    FluidStack get();
+
+    int getCapacity();
+}
